Moved email Subscription files to different packages, for convenience
