+++ src/main/java/com/example/harbour_swer/data/subscription/Subscription.java
@@ -1,4 +1,4 @@
-package com.example.harbour_swer.data;
+package com.example.harbour_swer.data.subscription;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
