+++ src/test/java/com/example/harbour_swer/controllers/EmailApiControllerTest.java
@@ -1,14 +1,14 @@
 package com.example.harbour_swer.controllers;
 
-import com.example.harbour_swer.data.Subscription;
-import com.example.harbour_swer.data.SubscriptionService;
+import com.example.harbour_swer.controllers.subscription_email.EmailApiController;
+import com.example.harbour_swer.data.subscription.Subscription;
+import com.example.harbour_swer.data.subscription.SubscriptionService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.ui.Model;
 
 import java.util.List;
